pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "synthlang"
version = "1.0.0"
description = "The Generative AI Pipeline DSL - Compose, evaluate, and deploy LLM pipelines with confidence"
readme = "README.md"
authors = [
    {name = "Michael Benjamin Crowe", email = "michael@crowelogic.com"}
]
license = {text = "Apache-2.0"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
]
keywords = ["llm", "ai", "pipeline", "dsl", "generative-ai", "orchestration", "langchain", "prompt-engineering"]
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "requests>=2.28",
    "rich>=13.0",
    "aiohttp>=3.8",
    "pydantic>=2.0",
]
requires-python = ">=3.8"

[project.urls]
Homepage = "https://github.com/MichaelCrowe11/synthlang"
Documentation = "https://synthlang.ai/docs"
Repository = "https://github.com/MichaelCrowe11/synthlang"
Issues = "https://github.com/MichaelCrowe11/synthlang/issues"

[project.scripts]
synthlang = "synthlang.cli:main"
synth = "synthlang.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "black>=23.0",
    "mypy>=1.0",
    "ruff>=0.1.0",
]
ide = [
    "flask>=2.3",
    "flask-cors>=4.0",
    "websockets>=11.0",
]
monitoring = [
    "prometheus-client>=0.16",
    "grafana-api>=1.0",
]

[tool.setuptools]
packages = ["synthlang"]

[tool.setuptools.package-data]
synthlang = ["*.yaml", "*.json", "templates/*", "static/*"]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py38"
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true